Refactor: 테스트 코드 디렉토리 구조 정리 및 리팩터링
- 테스트 코드를 API, 도메인, 통합 테스트로 구분하여 디렉토리 구조 정리
  - /api: Controller 관련 테스트 코드 분리
  - /application: Facade 및 Service 관련 테스트 코드 정리
  - /domain: 엔티티 및 비즈니스 로직 테스트 코드 정리
  - /integration: 통합 테스트를 위한 디렉토리 생성 및 테스트 코드 이동
- 테스트 코드의 가독성과 유지보수성 향상을 위해 구조 정리 완료
- 동시성 테스트와 단위 테스트를 각각 목적에 맞게 배치
